extract section and simplify ui page
