--- apps/website/components/PopoverSection/index.ts
@@ -1,1 +1,0 @@
-export * from "./PopoverSection";
