--- apps/website/components/Section/index.ts
@@ -1,1 +1,0 @@
-export * from "./Section";
