+++ apps/website/components/index.ts
@@ -1,3 +1,4 @@
-export * from "./PopoverSection";
 export * from "./Section";
 export * from "./ThemeSwitch";
+export * from "./SelectProp";
+export * from "./Card";
